Version 1.1.0: Fix crash on some devices. Add a Settings menu
Set renderscriptTargetApi to 18 instead of 19 to fix some crashes
with some decices.
Add a Settings menu. The settings icon is invisible by default.
Need to tap on the top right corner to display the icon, and then
tap on the icon within 4 seconds. The Settings menu lets the user
set the number of particles, the size of the particles, the number
of attration points, background color the particles colors, and
the force coefficients.